% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubic_cg.R
\name{hess_vec_fd}
\alias{hess_vec_fd}
\title{Create Hessian-Vector Product Function from Full Matrix}
\usage{
hess_vec_fd(H)
}
\arguments{
\item{H}{Hessian matrix (n x n)}
}
\value{
Function that computes H \%*\% v for any vector v
}
\description{
Wraps a full Hessian matrix in a function interface for use with
solve_cubic_cg.
}
\keyword{internal}
