% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qn_updates.R
\name{lsr1_multiply}
\alias{lsr1_multiply}
\title{L-SR1 Hessian-Vector Product}
\usage{
lsr1_multiply(history, v)
}
\arguments{
\item{history}{L-SR1 history list from \code{update_lsr1}.}

\item{v}{Vector to multiply by B.}
}
\value{
The product B*v as a numeric vector.
}
\description{
Computes the product B*v using the L-SR1 compact representation.
}
\details{
The L-SR1 Hessian has the compact form:
\deqn{B = \gamma I + \Psi (\Psi^T S)^{-1} \Psi^T}
where \eqn{\Psi = Y - \gamma S} and S, Y are matrices of stored vectors.

This function computes B*v in O(m^2 + mn) operations.
}
\keyword{internal}
